+++ common-db/src/main/java/cn/featherfly/common/db/mapping/JdbcMappingFactory.java
@@ -9,6 +9,7 @@ import cn.featherfly.common.repository.mapping.MappingFactory;
  * <p>
  * JdbcMappingFactory
  * </p>
+ * .
  *
  * @author zhongj
  */
@@ -27,4 +28,11 @@ public interface JdbcMappingFactory extends MappingFactory {
      * @return metadata
      */
     DatabaseMetadata getMetadata();
+
+    /**
+     * Gets the manager.
+     *
+     * @return the manager
+     */
+    SqlTypeMappingManager getSqlTypeMappingManager();
 }